created chips.module.ts for the ngModule and added forRoot() method
