+++ src/app/components/components/components.module.ts
@@ -45,7 +45,7 @@ import { CovalentChipsModule } from '../../../platform/chips';
     CovalentHighlightModule,
     CovalentMarkdownModule,
     CovalentJsonFormatterModule,
-    CovalentChipsModule,
+    CovalentChipsModule.forRoot(),
     componentsRoutes,
   ],
 })
